// src/Types/Exercise/Muscle.ts
export enum Muscle {
  Abdominals = "abdominals",
  Abductors = "abductors",
  Adductors = "adductors",
  Biceps = "biceps",
  Calves = "calves",
  Chest = "chest",
  Forearms = "forearms",
  Glutes = "glutes",
  Hamstrings = "hamstrings",
  Lats = "lats",
  LowerBack = "lower_back",
  MiddleBack = "middle_back",
  Neck = "neck",
  Quadriceps = "quadriceps",
  Traps = "traps",
  Triceps = "triceps"
}

export const muscles: Muscle[] = [
  Muscle.Abdominals,
  Muscle.Abductors,
  Muscle.Adductors,
  Muscle.Biceps,
  Muscle.Calves,
  Muscle.Chest,
  Muscle.Forearms,
  Muscle.Glutes,
  Muscle.Hamstrings,
  Muscle.Lats,
  Muscle.LowerBack,
  Muscle.MiddleBack,
  Muscle.Neck,
  Muscle.Quadriceps,
  Muscle.Traps,
  Muscle.Triceps
]